# topdown: cache undefined rule evaluations