--- a/topdown/cache.go
+++ b/topdown/cache.go
@@ -33,12 +33,13 @@ func (c *virtualCache) Pop() {
 	c.stack = c.stack[:len(c.stack)-1]
 }
 
-// Returns the resovled value of the AST term and a flag indicating if the value
+// Returns the resolved value of the AST term and a flag indicating if the value
 // should be interpretted as undefined:
-//    nil, true indicates the ref is undefined
-//    ast.Term, false indicates the ref is defined
-//    nil, false indicates the ref has not been cached
-//    ast.Term, true is impossible
+//
+//	nil, true indicates the ref is undefined
+//	ast.Term, false indicates the ref is defined
+//	nil, false indicates the ref has not been cached
+//	ast.Term, true is impossible
 func (c *virtualCache) Get(ref ast.Ref) (*ast.Term, bool) {
 	node := c.stack[len(c.stack)-1]
 	for i := 0; i < len(ref); i++ {
@@ -56,7 +57,7 @@ func (c *virtualCache) Get(ref ast.Ref) (*ast.Term, bool) {
 }
 
 // If value is a nil pointer, set the 'undefined' flag on the cache element to
-// indicate that the Ref has resovled to undefined.
+// indicate that the Ref has resolved to undefined.
 func (c *virtualCache) Put(ref ast.Ref, value *ast.Term) {
 	node := c.stack[len(c.stack)-1]
 	for i := 0; i < len(ref); i++ {