--- a/topdown/cache_test.go
+++ b/topdown/cache_test.go
@@ -14,7 +14,7 @@ func TestVirtualCacheCompositeKey(t *testing.T) {
 	cache := newVirtualCache()
 	ref := ast.MustParseRef("data.x.y[[1]].z")
 	cache.Put(ref, ast.BooleanTerm(true))
-	result := cache.Get(ref)
+	result, _ := cache.Get(ref)
 	if !result.Equal(ast.BooleanTerm(true)) {
 		t.Fatalf("Expected true but got %v", result)
 	}
@@ -25,7 +25,7 @@ func TestVirtualCacheInvalidate(t *testing.T) {
 	cache.Push()
 	cache.Put(ast.MustParseRef("data.x.p"), ast.BooleanTerm(true))
 	cache.Pop()
-	result := cache.Get(ast.MustParseRef("data.x.p"))
+	result, _ := cache.Get(ast.MustParseRef("data.x.p"))
 	if result != nil {
 		t.Fatal("Expected nil result but got:", result)
 	}