In the update from 0.45 -> 0.46 I found the new Ref Head feature dramatically reduced the performance of my policy bundles. The issue came from two sets of rules in the following format:

main.rego
```rego
package main

import data.loader

partial_rule[key] := value {
  value := loader[key]
}  
```

loader.rego
```rego
package loader

conditional_not_met {
  input.a == 2
}

conditional_met {
  input.a == 1
}

partial_object["some_key"] := result {
  conditional_not_met
  result := "value"
}

partial_object["some_other_key"] := result {
  conditional_met
  result := "value"
}
```

In OPA 0.45, the result of `partial_object` would be cached under `data.loader.partial_object` so when the `partial_rule` from `main.rego` is evaluated it loads the cached result of `partial_object`. In OPA 0.46 and greater, the Ref Head system means that `partial_rule` iterates over `data.loader.partial_object.some_key`, `data.loader.partial_object.some_other_key`, etc. Because rules that evaluate to `undefined` are not cached, this causes OPA to re-evaluate the conditional rules multiple times depending on how these rules are referenced, which resulted in the performance regression I observed with my Rego bundle. 

This change creates a mechanism to cache an 'undefined' result in the virtualCache during topdown evaluation. This prevents rules from being re-run when they are evaluated to 'undefined', which fixes my observed performance regression and then some. 

I added a benchmark for the specific pattern of my bundle's rules and observe the following results: 
```
name                                  old time/op  new time/op  delta
PartialObjectRuleCrossModule/10-12     311µs ±10%   289µs ± 8%     ~     (p=0.095 n=5+5)
PartialObjectRuleCrossModule/100-12   2.11ms ± 1%  1.82ms ± 2%  -13.80%  (p=0.008 n=5+5)
PartialObjectRuleCrossModule/1000-12  25.2ms ± 7%  20.5ms ± 2%  -18.70%  (p=0.008 n=5+5)
```

And profiling my bundle with `--count=100` I see the following results:

OPA 0.47:
```
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+
|             METRIC             |    MIN    |    MAX     |       MEAN       |          90%           |          99%           |
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+
| timer_rego_external_resolve_ns | 1176      | 9685       | 1778.69          | 2234.9                 | 9627.40999999997       |
| timer_rego_load_bundles_ns     | 44225828  | 103708720  | 6.263822416e+07  | 7.609671660000001e+07  | 1.0362247983999996e+08 |
| timer_rego_module_compile_ns   | 512578468 | 1085634008 | 6.3867703977e+08 | 7.768929584000001e+08  | 1.0848653646999996e+09 |
| timer_rego_module_parse_ns     | 41457186  | 91070535   | 5.850644291e+07  | 7.27230553e+07         | 9.106606514999999e+07  |
| timer_rego_query_compile_ns    | 194742    | 648664     | 271893.74        | 341785.4               | 646902.479999999       |
| timer_rego_query_eval_ns       | 9821426   | 34616892   | 1.500954556e+07  | 2.1171452800000004e+07 | 3.456969626999997e+07  |
| timer_rego_query_parse_ns      | 46020     | 240162     | 68757.42         | 101263                 | 239167.0799999995      |
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+
+-----------+-------------+-----------+------------+-------------+----------+----------+
|    MIN    |     MAX     |   MEAN    |    90%     |     99%     | NUM EVAL | NUM REDO |
+-----------+-------------+-----------+------------+-------------+----------+----------+
| 251.572µs | 11.189607ms | 910.492µs | 1.110284ms | 11.159263ms | 164      | 123      |
| 246.756µs | 1.707731ms  | 397.249µs | 503.186µs  | 1.705948ms  | 60       | 45       |
| 184.992µs | 1.678417ms  | 322.281µs | 435.221µs  | 1.668585ms  | 44       | 33       |
| 165.177µs | 937.035µs   | 276.863µs | 375.635µs  | 934.507µs   | 1        | 0        |
| 140.245µs | 923.027µs   | 242.516µs | 349.156µs  | 918.391µs   | 1        | 51       |
| 122.434µs | 780.667µs   | 207.106µs | 301.478µs  | 777.217µs   | 33       | 33       |
| 89.461µs  | 739.014µs   | 177.825µs | 247.76µs   | 735.171µs   | 9        | 6        |
| 83.144µs  | 534.366µs   | 156.152µs | 243.657µs  | 532.373µs   | 34       | 20       |
| 78.456µs  | 503.971µs   | 142.514µs | 210.178µs  | 502.253µs   | 2        | 0        |
| 76.993µs  | 475.92µs    | 131.156µs | 197.532µs  | 474.113µs   | 8        | 6        |
+-----------+-------------+-----------+------------+-------------+----------+----------+

```

OPA HEAD + this patch:
```
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+
|             METRIC             |    MIN    |    MAX     |       MEAN       |          90%           |          99%           |
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+
| timer_rego_external_resolve_ns | 1132      | 35263      | 2183.36          | 2448.3                 | 34998.53999999986      |
| timer_rego_load_bundles_ns     | 46459458  | 100135005  | 6.176983085e+07  | 7.873112670000002e+07  | 1.0005333212999997e+08 |
| timer_rego_module_compile_ns   | 454740302 | 1038241310 | 6.1283400342e+08 | 7.755072528000003e+08  | 1.0378503374099998e+09 |
| timer_rego_module_parse_ns     | 42972248  | 96072401   | 5.759438505e+07  | 7.25270605e+07         | 9.598284676999995e+07  |
| timer_rego_query_compile_ns    | 176852    | 716035     | 290369.48        | 424998.4000000001      | 715368.6099999996      |
| timer_rego_query_eval_ns       | 6804308   | 54873736   | 1.13959951e+07   | 1.4939162700000001e+07 | 5.46777581699999e+07   |
| timer_rego_query_parse_ns      | 44497     | 185458     | 62159.75         | 86102.70000000003      | 185112.00999999983     |
+--------------------------------+-----------+------------+------------------+------------------------+------------------------+

+-----------+-------------+-----------+-----------+-------------+----------+----------+
|    MIN    |     MAX     |   MEAN    |    90%    |     99%     | NUM EVAL | NUM REDO |
+-----------+-------------+-----------+-----------+-------------+----------+----------+
| 172.479µs | 14.888451ms | 776.03µs  | 530.996µs | 14.829465ms | 1        | 0        |
| 157.932µs | 12.122327ms | 384.667µs | 382.612µs | 12.011663ms | 1        | 1        |
| 138.617µs | 10.583771ms | 324.811µs | 301.466µs | 10.482737ms | 1        | 51       |
| 76.402µs  | 963.717µs   | 165.296µs | 247.404µs | 958.578µs   | 2        | 2        |
| 67.489µs  | 529.998µs   | 132.783µs | 208.574µs | 528.853µs   | 1        | 1        |
| 62.968µs  | 395.385µs   | 116.032µs | 161.824µs | 395.314µs   | 1        | 0        |
| 60.333µs  | 379.738µs   | 106.442µs | 149.237µs | 379.669µs   | 1        | 0        |
| 56.969µs  | 349.754µs   | 98.16µs   | 130.82µs  | 348.853µs   | 4        | 3        |
| 56.201µs  | 332.023µs   | 92.503µs  | 125.952µs | 330.703µs   | 16       | 16       |
| 52.208µs  | 326.965µs   | 88.236µs  | 118.56µs  | 325.683µs   | 1        | 0        |
+-----------+-------------+-----------+-----------+-------------+----------+----------+
```

I did not observe any test regressions when making this change.