@edpaget 2023-01-04
Thanks for the review @srenatus! 

I looked at the output of explain=full format=pretty and when it hits a cached undefined value I see something like 

```
path:35                           | | Index rule_name (matched 5 rules)
```

Which I think is consistent with when a cached value is hit that isn't undefined. I also added the cache hit metric increment in the branch for a undefined result from the cache.