@srenatus 2023-01-05
I'm inclined to merge this, but I'd like to have another short head-to-head with the code tomorrow. Thanks again for contributing, this is good stuff.