@srenatus 2023-01-06
✔️ Cache hit metric works as expected, for x.rego
```rego
package p
x { false }
y { x }
y { x }
```

```
$ opa eval -fpretty --explain=debug -d x.rego data.p.y --instrument | grep cache
| counter_eval_op_virtual_cache_hit                        | 1                  |
| counter_eval_op_virtual_cache_miss                       | 2                  |
$ opa-0.46.1 eval -fpretty --explain=debug -d x.rego data.p.y --instrument | grep cache 
| counter_eval_op_virtual_cache_miss                       | 3                  |
```

✔️ Spot-checking `with`, it also all works as expected,
```rego
package p
x { input }
y { x }
y { x with input as true }
```

```
$ opa eval -fpretty --explain=debug -d x.rego data.p.y --instrument  | grep cache
| counter_eval_op_virtual_cache_miss                       | 3                  |
```